tests: update requirements
This project use gitlab-ci and don't need any more coveralls.

With DNS verification we don't need a FUSE server and we need
dnspython3 package.

The dnspython has merged code for python2 and python3, but the
current release has an error with TSIG key management (due to the
code merge). So we are now stuck with the dnspython3 package.